backend/requirements.txt
werkzeug==2.0.3
flask==2.0.1
flask-cors==3.0.10
flask-limiter==3.5.0
reportlab==4.3.1
Pillow==11.0.0
langchain==0.1.9
zhipuai==2.0.1
python-dotenv==1.0.0
tenacity==8.2.3 